[front/components/agent_builder] - feature: implement data source view filtering in agent builder context
 - Include useMemo to optimize filtering supported data sources by feature flags
 - Replace direct console.warn with centralized logger for invalid page IDs in AddSearchSheet component
 - Exclude table-only data sources from document search feature within the agent builder
 - Use context to pass down pre-filtered supportedDataSourceViews to child components
