+++ front/components/agent_builder/AgentBuilderContext.tsx
@@ -1,18 +1,23 @@
-import { createContext, useContext, useState } from "react";
+import { createContext, useContext, useMemo, useState } from "react";
 import { useEffect } from "react";
 
 import { MCPServerViewsProvider } from "@app/components/assistant_builder/contexts/MCPServerViewsContext";
 import { SpacesProvider } from "@app/components/assistant_builder/contexts/SpacesContext";
-import type { WorkspaceType } from "@app/types";
+import { supportsDocumentsData } from "@app/lib/data_sources";
+import { useDataSourceViews } from "@app/lib/swr/data_source_views";
+import { useFeatureFlags } from "@app/lib/swr/workspaces";
+import type { DataSourceViewType, WorkspaceType } from "@app/types";
 
 type AgentBuilderContextType = {
   owner: WorkspaceType;
+  supportedDataSourceViews: DataSourceViewType[];
   isPreviewPanelOpen: boolean;
   setIsPreviewPanelOpen: React.Dispatch<React.SetStateAction<boolean>>;
 };
 
 export const AgentBuilderContext = createContext<AgentBuilderContextType>({
   owner: {} as WorkspaceType,
+  supportedDataSourceViews: [],
   isPreviewPanelOpen: true,
   setIsPreviewPanelOpen: () => {},
 });
@@ -20,7 +25,7 @@ export const AgentBuilderContext = createContext<AgentBuilderContextType>({
 interface AgentBuilderContextProps
   extends Omit<
     AgentBuilderContextType,
-    "isPreviewPanelOpen" | "setIsPreviewPanelOpen"
+    "supportedDataSourceViews" | "isPreviewPanelOpen" | "setIsPreviewPanelOpen"
   > {
   children: React.ReactNode;
 }
@@ -30,6 +35,18 @@ export function AgentBuilderProvider({
   children,
 }: AgentBuilderContextProps) {
   const [isPreviewPanelOpen, setIsPreviewPanelOpen] = useState(false);
+
+  const { dataSourceViews } = useDataSourceViews(owner);
+  const { featureFlags } = useFeatureFlags({
+    workspaceId: owner.sId,
+  });
+
+  // Filter data sources for document search (excluding table-only sources)
+  const supportedDataSourceViews = useMemo(() => {
+    return dataSourceViews.filter((dsv) =>
+      supportsDocumentsData(dsv.dataSource, featureFlags)
+    );
+  }, [dataSourceViews, featureFlags]);
 
   useEffect(() => {
     const mediaQuery = window.matchMedia("(min-width: 1024px)");
@@ -49,6 +66,7 @@ export function AgentBuilderProvider({
     <AgentBuilderContext.Provider
       value={{
         owner,
+        supportedDataSourceViews,
         isPreviewPanelOpen,
         setIsPreviewPanelOpen,
       }}
